refactor: migrate apply delay to transaction
